+++ src/dms/backend/EdFi.DataManagementService.Backend.Mssql.Tests.Integration/MssqlDocumentCacheMaterializerCoherenceTests.cs
@@ -9,6 +9,7 @@ using EdFi.DataManagementService.Backend.Etag;
 using EdFi.DataManagementService.Backend.External;
 using EdFi.DataManagementService.Backend.External.Plans;
 using EdFi.DataManagementService.Backend.Mssql;
+using EdFi.DataManagementService.Backend.Tests.Common;
 using EdFi.DataManagementService.Backend.Tests.Integration.Common;
 using EdFi.DataManagementService.Core.External.Model;
 using FluentAssertions;
@@ -24,12 +25,9 @@ namespace EdFi.DataManagementService.Backend.Mssql.Tests.Integration;
 public class Given_Mssql_DocumentCacheMaterializer_Coherence
 {
     private const long DocumentId = 980101;
-    private const short ResourceKeyId = 11;
     private const long ContentVersion = 222;
     private static readonly Guid DocumentGuid = Guid.Parse("aaaaaaaa-bbbb-cccc-dddd-000000009801");
     private static readonly DateTimeOffset LastModifiedAt = new(2026, 7, 30, 14, 15, 16, TimeSpan.Zero);
-    private static readonly QualifiedResourceName SchoolResource = new("Ed-Fi", "School");
-
     private string _databaseName = null!;
     private string _connectionString = null!;
     private MappingSet _mappingSet = null!;
@@ -44,7 +42,7 @@ public class Given_Mssql_DocumentCacheMaterializer_Coherence
         _databaseName = MssqlTestDatabaseHelper.GenerateUniqueDatabaseName();
         MssqlTestDatabaseHelper.CreateDatabase(_databaseName);
         _connectionString = MssqlTestDatabaseHelper.BuildConnectionString(_databaseName);
-        _mappingSet = CreateMappingSet(SqlDialect.Mssql);
+        _mappingSet = DocumentCacheMaterializerCoherenceMappingSet.Create(SqlDialect.Mssql);
 
         await using var connection = new SqlConnection(_connectionString);
         await connection.OpenAsync();
@@ -157,7 +155,7 @@ public class Given_Mssql_DocumentCacheMaterializer_Coherence
         var materializationDataStore = new AmbientDocumentCacheMaterializationDataStore(
             commandExecutor,
             new MutatingDocumentHydrator(
-                _mappingSet.ReadPlansByResource[SchoolResource],
+                _mappingSet.ReadPlansByResource[DocumentCacheMaterializerCoherenceMappingSet.SchoolResource],
                 mutateDuringHydration
             )
         );
@@ -183,102 +181,6 @@ public class Given_Mssql_DocumentCacheMaterializer_Coherence
             DocumentCacheMaterializationPurpose.Fixture,
             CancellationToken.None
         );
-
-    private static MappingSet CreateMappingSet(SqlDialect dialect)
-    {
-        var readPlan = CreateReadPlan(dialect);
-        var resourceKey = new ResourceKeyEntry(ResourceKeyId, SchoolResource, "1.0", false);
-        var concreteResourceModel = new ConcreteResourceModel(
-            resourceKey,
-            ResourceStorageKind.RelationalTables,
-            readPlan.Model
-        );
-        var effectiveSchema = new EffectiveSchemaInfo(
-            ApiSchemaFormatVersion: "1.0",
-            RelationalMappingVersion: "v1",
-            EffectiveSchemaHash: "0123456789abcdef0123456789abcdef0123456789abcdef0123456789abcdef",
-            ResourceKeyCount: 1,
-            ResourceKeySeedHash: new byte[32],
-            SchemaComponentsInEndpointOrder: [],
-            ResourceKeysInIdOrder: [resourceKey]
-        );
-
-        return new MappingSet(
-            new MappingSetKey(effectiveSchema.EffectiveSchemaHash, dialect, "v1"),
-            new DerivedRelationalModelSet(
-                effectiveSchema,
-                dialect,
-                ProjectSchemasInEndpointOrder: [],
-                ConcreteResourcesInNameOrder: [concreteResourceModel],
-                AbstractIdentityTablesInNameOrder: [],
-                AbstractUnionViewsInNameOrder: [],
-                IndexesInCreateOrder: [],
-                TriggersInCreateOrder: []
-            ),
-            WritePlansByResource: new Dictionary<QualifiedResourceName, ResourceWritePlan>(),
-            ReadPlansByResource: new Dictionary<QualifiedResourceName, ResourceReadPlan>
-            {
-                [SchoolResource] = readPlan,
-            },
-            ResourceKeyIdByResource: new Dictionary<QualifiedResourceName, short>
-            {
-                [SchoolResource] = ResourceKeyId,
-            },
-            ResourceKeyById: new Dictionary<short, ResourceKeyEntry> { [ResourceKeyId] = resourceKey },
-            SecurableElementColumnPathsByResource: new Dictionary<
-                QualifiedResourceName,
-                IReadOnlyList<ResolvedSecurableElementPath>
-            >()
-        );
-    }
-
-    private static ResourceReadPlan CreateReadPlan(SqlDialect dialect)
-    {
-        var rootTable = new DbTableModel(
-            new DbTableName(new DbSchemaName("edfi"), "School"),
-            new JsonPathExpression("$", []),
-            new TableKey(
-                "PK_School",
-                [new DbKeyColumn(new DbColumnName("DocumentId"), ColumnKind.ParentKeyPart)]
-            ),
-            [
-                new DbColumnModel(
-                    new DbColumnName("DocumentId"),
-                    ColumnKind.ParentKeyPart,
-                    new RelationalScalarType(ScalarKind.Int64),
-                    false,
-                    null,
-                    null
-                ),
-            ],
-            []
-        )
-        {
-            IdentityMetadata = new DbTableIdentityMetadata(
-                DbTableKind.Root,
-                [new DbColumnName("DocumentId")],
-                [new DbColumnName("DocumentId")],
-                [],
-                []
-            ),
-        };
-
-        return new ResourceReadPlan(
-            new RelationalResourceModel(
-                SchoolResource,
-                new DbSchemaName("edfi"),
-                ResourceStorageKind.RelationalTables,
-                rootTable,
-                [rootTable],
-                [],
-                []
-            ),
-            KeysetTableConventions.GetKeysetTableContract(dialect),
-            [new TableReadPlan(rootTable, "select DocumentId")],
-            [],
-            []
-        );
-    }
 
     private static async Task ExecuteSql(SqlConnection connection, string sql)
     {
