+++ src/dms/backend/EdFi.DataManagementService.Backend.Postgresql.Tests.Integration/PostgresqlDocumentCacheMaterializerCoherenceTests.cs
@@ -9,6 +9,7 @@ using EdFi.DataManagementService.Backend.Etag;
 using EdFi.DataManagementService.Backend.External;
 using EdFi.DataManagementService.Backend.External.Plans;
 using EdFi.DataManagementService.Backend.Postgresql;
+using EdFi.DataManagementService.Backend.Tests.Common;
 using EdFi.DataManagementService.Backend.Tests.Integration.Common;
 using EdFi.DataManagementService.Core.External.Model;
 using FluentAssertions;
@@ -23,12 +24,9 @@ namespace EdFi.DataManagementService.Backend.Postgresql.Tests.Integration;
 public class Given_Postgresql_DocumentCacheMaterializer_Coherence
 {
     private const long DocumentId = 980101;
-    private const short ResourceKeyId = 11;
     private const long ContentVersion = 222;
     private static readonly Guid DocumentGuid = Guid.Parse("aaaaaaaa-bbbb-cccc-dddd-000000009801");
     private static readonly DateTimeOffset LastModifiedAt = new(2026, 7, 30, 14, 15, 16, TimeSpan.Zero);
-    private static readonly QualifiedResourceName SchoolResource = new("Ed-Fi", "School");
-
     private PostgresqlGeneratedDdlTestDatabase _database = null!;
     private NpgsqlDataSource _dataSource = null!;
     private MappingSet _mappingSet = null!;
@@ -38,7 +36,7 @@ public class Given_Postgresql_DocumentCacheMaterializer_Coherence
     {
         _database = await PostgresqlGeneratedDdlTestDatabase.CreateEmptyAsync();
         _dataSource = NpgsqlDataSource.Create(_database.ConnectionString);
-        _mappingSet = CreateMappingSet(SqlDialect.Pgsql);
+        _mappingSet = DocumentCacheMaterializerCoherenceMappingSet.Create(SqlDialect.Pgsql);
 
         await using var connection = await _dataSource.OpenConnectionAsync();
         await ExecuteSql(
@@ -148,7 +146,7 @@ public class Given_Postgresql_DocumentCacheMaterializer_Coherence
         var materializationDataStore = new AmbientDocumentCacheMaterializationDataStore(
             commandExecutor,
             new MutatingDocumentHydrator(
-                _mappingSet.ReadPlansByResource[SchoolResource],
+                _mappingSet.ReadPlansByResource[DocumentCacheMaterializerCoherenceMappingSet.SchoolResource],
                 mutateDuringHydration
             )
         );
@@ -174,102 +172,6 @@ public class Given_Postgresql_DocumentCacheMaterializer_Coherence
             DocumentCacheMaterializationPurpose.Fixture,
             CancellationToken.None
         );
-
-    private static MappingSet CreateMappingSet(SqlDialect dialect)
-    {
-        var readPlan = CreateReadPlan(dialect);
-        var resourceKey = new ResourceKeyEntry(ResourceKeyId, SchoolResource, "1.0", false);
-        var concreteResourceModel = new ConcreteResourceModel(
-            resourceKey,
-            ResourceStorageKind.RelationalTables,
-            readPlan.Model
-        );
-        var effectiveSchema = new EffectiveSchemaInfo(
-            ApiSchemaFormatVersion: "1.0",
-            RelationalMappingVersion: "v1",
-            EffectiveSchemaHash: "0123456789abcdef0123456789abcdef0123456789abcdef0123456789abcdef",
-            ResourceKeyCount: 1,
-            ResourceKeySeedHash: new byte[32],
-            SchemaComponentsInEndpointOrder: [],
-            ResourceKeysInIdOrder: [resourceKey]
-        );
-
-        return new MappingSet(
-            new MappingSetKey(effectiveSchema.EffectiveSchemaHash, dialect, "v1"),
-            new DerivedRelationalModelSet(
-                effectiveSchema,
-                dialect,
-                ProjectSchemasInEndpointOrder: [],
-                ConcreteResourcesInNameOrder: [concreteResourceModel],
-                AbstractIdentityTablesInNameOrder: [],
-                AbstractUnionViewsInNameOrder: [],
-                IndexesInCreateOrder: [],
-                TriggersInCreateOrder: []
-            ),
-            WritePlansByResource: new Dictionary<QualifiedResourceName, ResourceWritePlan>(),
-            ReadPlansByResource: new Dictionary<QualifiedResourceName, ResourceReadPlan>
-            {
-                [SchoolResource] = readPlan,
-            },
-            ResourceKeyIdByResource: new Dictionary<QualifiedResourceName, short>
-            {
-                [SchoolResource] = ResourceKeyId,
-            },
-            ResourceKeyById: new Dictionary<short, ResourceKeyEntry> { [ResourceKeyId] = resourceKey },
-            SecurableElementColumnPathsByResource: new Dictionary<
-                QualifiedResourceName,
-                IReadOnlyList<ResolvedSecurableElementPath>
-            >()
-        );
-    }
-
-    private static ResourceReadPlan CreateReadPlan(SqlDialect dialect)
-    {
-        var rootTable = new DbTableModel(
-            new DbTableName(new DbSchemaName("edfi"), "School"),
-            new JsonPathExpression("$", []),
-            new TableKey(
-                "PK_School",
-                [new DbKeyColumn(new DbColumnName("DocumentId"), ColumnKind.ParentKeyPart)]
-            ),
-            [
-                new DbColumnModel(
-                    new DbColumnName("DocumentId"),
-                    ColumnKind.ParentKeyPart,
-                    new RelationalScalarType(ScalarKind.Int64),
-                    false,
-                    null,
-                    null
-                ),
-            ],
-            []
-        )
-        {
-            IdentityMetadata = new DbTableIdentityMetadata(
-                DbTableKind.Root,
-                [new DbColumnName("DocumentId")],
-                [new DbColumnName("DocumentId")],
-                [],
-                []
-            ),
-        };
-
-        return new ResourceReadPlan(
-            new RelationalResourceModel(
-                SchoolResource,
-                new DbSchemaName("edfi"),
-                ResourceStorageKind.RelationalTables,
-                rootTable,
-                [rootTable],
-                [],
-                []
-            ),
-            KeysetTableConventions.GetKeysetTableContract(dialect),
-            [new TableReadPlan(rootTable, "select DocumentId")],
-            [],
-            []
-        );
-    }
 
     private static async Task ExecuteSql(NpgsqlConnection connection, string sql)
     {
