+++ src/dms/backend/EdFi.DataManagementService.Backend.Postgresql/PostgresqlDocumentCacheMaterializationDataStore.cs
@@ -46,7 +46,7 @@ internal sealed class PostgresqlDocumentCacheMaterializationDataStore(
 
         _logger.LogDebug(
             "Executing PostgreSQL DocumentCache materialization command for target {TargetKey} with {ParameterCount} parameters",
-            request.TargetContext.TargetKey,
+            LogSanitizer.SanitizeForLog(request.TargetContext.TargetKey.ToString()),
             command.Parameters.Count
         );
