+++ src/dms/backend/EdFi.DataManagementService.Backend.Tests.Common/DocumentCacheMaterializerCoherenceMappingSet.cs
@@ -1,0 +1,112 @@
+// SPDX-License-Identifier: Apache-2.0
+// Licensed to the Ed-Fi Alliance under one or more agreements.
+// The Ed-Fi Alliance licenses this file to you under the Apache License, Version 2.0.
+// See the LICENSE and NOTICES files in the project root for more information.
+
+using EdFi.DataManagementService.Backend.External;
+using EdFi.DataManagementService.Backend.External.Plans;
+
+namespace EdFi.DataManagementService.Backend.Tests.Common;
+
+internal static class DocumentCacheMaterializerCoherenceMappingSet
+{
+    private const short ResourceKeyId = 11;
+
+    public static readonly QualifiedResourceName SchoolResource = new("Ed-Fi", "School");
+
+    public static MappingSet Create(SqlDialect dialect)
+    {
+        var readPlan = CreateReadPlan(dialect);
+        var resourceKey = new ResourceKeyEntry(ResourceKeyId, SchoolResource, "1.0", false);
+        var concreteResourceModel = new ConcreteResourceModel(
+            resourceKey,
+            ResourceStorageKind.RelationalTables,
+            readPlan.Model
+        );
+        var effectiveSchema = new EffectiveSchemaInfo(
+            ApiSchemaFormatVersion: "1.0",
+            RelationalMappingVersion: "v1",
+            EffectiveSchemaHash: "0123456789abcdef0123456789abcdef0123456789abcdef0123456789abcdef",
+            ResourceKeyCount: 1,
+            ResourceKeySeedHash: new byte[32],
+            SchemaComponentsInEndpointOrder: [],
+            ResourceKeysInIdOrder: [resourceKey]
+        );
+
+        return new MappingSet(
+            new MappingSetKey(effectiveSchema.EffectiveSchemaHash, dialect, "v1"),
+            new DerivedRelationalModelSet(
+                effectiveSchema,
+                dialect,
+                ProjectSchemasInEndpointOrder: [],
+                ConcreteResourcesInNameOrder: [concreteResourceModel],
+                AbstractIdentityTablesInNameOrder: [],
+                AbstractUnionViewsInNameOrder: [],
+                IndexesInCreateOrder: [],
+                TriggersInCreateOrder: []
+            ),
+            WritePlansByResource: new Dictionary<QualifiedResourceName, ResourceWritePlan>(),
+            ReadPlansByResource: new Dictionary<QualifiedResourceName, ResourceReadPlan>
+            {
+                [SchoolResource] = readPlan,
+            },
+            ResourceKeyIdByResource: new Dictionary<QualifiedResourceName, short>
+            {
+                [SchoolResource] = ResourceKeyId,
+            },
+            ResourceKeyById: new Dictionary<short, ResourceKeyEntry> { [ResourceKeyId] = resourceKey },
+            SecurableElementColumnPathsByResource: new Dictionary<
+                QualifiedResourceName,
+                IReadOnlyList<ResolvedSecurableElementPath>
+            >()
+        );
+    }
+
+    private static ResourceReadPlan CreateReadPlan(SqlDialect dialect)
+    {
+        var rootTable = new DbTableModel(
+            new DbTableName(new DbSchemaName("edfi"), "School"),
+            new JsonPathExpression("$", []),
+            new TableKey(
+                "PK_School",
+                [new DbKeyColumn(new DbColumnName("DocumentId"), ColumnKind.ParentKeyPart)]
+            ),
+            [
+                new DbColumnModel(
+                    new DbColumnName("DocumentId"),
+                    ColumnKind.ParentKeyPart,
+                    new RelationalScalarType(ScalarKind.Int64),
+                    false,
+                    null,
+                    null
+                ),
+            ],
+            []
+        )
+        {
+            IdentityMetadata = new DbTableIdentityMetadata(
+                DbTableKind.Root,
+                [new DbColumnName("DocumentId")],
+                [new DbColumnName("DocumentId")],
+                [],
+                []
+            ),
+        };
+
+        return new ResourceReadPlan(
+            new RelationalResourceModel(
+                SchoolResource,
+                new DbSchemaName("edfi"),
+                ResourceStorageKind.RelationalTables,
+                rootTable,
+                [rootTable],
+                [],
+                []
+            ),
+            KeysetTableConventions.GetKeysetTableContract(dialect),
+            [new TableReadPlan(rootTable, "select DocumentId")],
+            [],
+            []
+        );
+    }
+}
