+++ src/dms/backend/Fixtures/document-cache/materialized-documents/extension-student-school-association/expected-cache-row.json
@@ -5,7 +5,7 @@
   "resourceName": "StudentSchoolAssociation",
   "resourceVersion": "5.2.0",
   "contentVersion": 333,
-  "lastModifiedAt": "2026-07-31T10:11:12.9876543Z",
+  "lastModifiedAt": "2026-07-31T10:11:12.987654Z",
   "streamEtag": "333-53ba4ec6.j._.l.i",
   "documentJson": {
     "id": "aaaaaaaa-bbbb-cccc-dddd-000000001101",
