+++ src/dms/backend/Fixtures/document-cache/materialized-documents/extension-student-school-association/source-setup.json
@@ -5,7 +5,7 @@
       "documentUuid": "aaaaaaaa-bbbb-cccc-dddd-000000001101",
       "resourceKeyId": 310,
       "contentVersion": 333,
-      "contentLastModifiedAt": "2026-07-31T10:11:12.9876543Z"
+      "contentLastModifiedAt": "2026-07-31T10:11:12.987654Z"
     },
     {
       "documentId": 971201,
