+++ src/dms/backend/Fixtures/document-cache/materialized-documents/layout-contract-school/expected-cache-row.json
@@ -5,8 +5,8 @@
   "resourceName": "School",
   "resourceVersion": "5.3.0",
   "contentVersion": 42,
-  "lastModifiedAt": "2026-01-15T17:18:19.1234567Z",
-  "streamEtag": "\"42-cache-projection\"",
+  "lastModifiedAt": "2026-01-15T17:18:19.123456Z",
+  "streamEtag": "42-00000000.j._.l.i",
   "documentJson": {
     "id": "11111111-1111-1111-1111-111111111111",
     "_lastModifiedDate": "2026-01-15T17:18:19Z",
