+++ src/dms/backend/Fixtures/document-cache/materialized-documents/layout-contract-school/expected-public-cdc-document.json
@@ -4,6 +4,6 @@
     "_lastModifiedDate": "2026-01-15T17:18:19Z",
     "schoolId": 255901001,
     "nameOfInstitution": "Layout Contract School",
-    "_etag": "\"42-cache-projection\""
+    "_etag": "42-00000000.j._.l.i"
   }
 }
