+++ src/dms/backend/Fixtures/document-cache/materialized-documents/layout-contract-school/expected-stream-etag.json
@@ -1,3 +1,3 @@
 {
-  "streamEtag": "\"42-cache-projection\""
+  "streamEtag": "42-00000000.j._.l.i"
 }
