+++ src/dms/backend/Fixtures/document-cache/materialized-documents/layout-contract-school/source-setup.json
@@ -5,7 +5,7 @@
       "documentUuid": "11111111-1111-1111-1111-111111111111",
       "resourceKeyId": 1,
       "contentVersion": 42,
-      "contentLastModifiedAt": "2026-01-15T17:18:19.1234567Z"
+      "contentLastModifiedAt": "2026-01-15T17:18:19.123456Z"
     }
   ],
   "descriptors": [],
