+++ src/dms/backend/Fixtures/document-cache/materialized-documents/ordinary-link-bearing-student-school-association/expected-cache-row.json
@@ -5,7 +5,7 @@
   "resourceName": "StudentSchoolAssociation",
   "resourceVersion": "1.0",
   "contentVersion": 222,
-  "lastModifiedAt": "2026-07-30T14:15:16.1234567Z",
+  "lastModifiedAt": "2026-07-30T14:15:16.123456Z",
   "streamEtag": "222-01234567.j._.l.i",
   "documentJson": {
     "id": "aaaaaaaa-bbbb-cccc-dddd-000000000101",
