+++ src/dms/backend/Fixtures/document-cache/materialized-documents/ordinary-link-bearing-student-school-association/source-setup.json
@@ -5,7 +5,7 @@
       "documentUuid": "aaaaaaaa-bbbb-cccc-dddd-000000000101",
       "resourceKeyId": 11,
       "contentVersion": 222,
-      "contentLastModifiedAt": "2026-07-30T14:15:16.1234567Z"
+      "contentLastModifiedAt": "2026-07-30T14:15:16.123456Z"
     },
     {
       "documentId": 970201,
